Update storage temperature field type and sync logic
Changed `storage_temperature` field type from `Float` to `Integer` in `product_template.py`. Updated test cases and product data transformation logic to handle the integer storage temperature values accordingly.

+++ product_food_fields/models/product_template.py
@@ -8,5 +8,5 @@ class ProductTemplate(models.Model):
     origin = fields.Char(help="Land where the product is produced, an ISO 3166 code")
     usage_tips = fields.Char(help="Usage tips")
     days_until_expiry = fields.Integer(help="Days until expiration")
-    storage_temperature = fields.Float(help="Storage temperature in Celcius")
+    storage_temperature = fields.Integer(help="Storage temperature in Celcius")
     product_quality_id = fields.Many2one("product_food_fields.product_quality", string="Product quality")
